we are so back

+++ src/io/utils/unseekable.rs
@@ -2,7 +2,7 @@ use std::io::{self, Read, Seek, SeekFrom, Write};
 
 pub struct Unseekable<T> {
     inner: T,
-    i: usize,
+    i: u64,
 }
 
 impl<T> Unseekable<T> {
@@ -27,7 +27,7 @@ impl<T: Read> Read for Unseekable<T> {
     fn read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
         let res = self.inner.read(buf);
         if let Ok(n) = res {
-            self.i += n;
+            self.i += n as u64;
         }
 
         res
@@ -38,7 +38,7 @@ impl<T: Write> Write for Unseekable<T> {
     fn write(&mut self, buf: &[u8]) -> io::Result<usize> {
         let res = self.inner.write(buf);
         if let Ok(n) = res {
-            self.i += n;
+            self.i += n as u64;
         }
 
         res
@@ -52,7 +52,7 @@ impl<T: Write> Write for Unseekable<T> {
 impl<T> Seek for Unseekable<T> {
     fn seek(&mut self, offset: SeekFrom) -> io::Result<u64> {
         match offset {
-            SeekFrom::Current(0) => Ok(self.i as u64),
+            SeekFrom::Current(0) => Ok(self.i),
             _ => Err(io::ErrorKind::Unsupported.into()),
         }
     }
